requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
psycopg[binary]>=3.2,<4.0
supabase>=2.6,<3.0
PyJWT>=2.8,<3.0
requests>=2.32,<3.0
